==> playground/01_extract_paragraphs.ipynb <==
import pandas as pd

SOURCE = '../data/kilt_wiki_small.jsonl'

df = pd.read_json(SOURCE, lines=True)
df

df['paragraphs'] = df['text'].apply(lambda t: t['paragraph'])

df.loc[df['wikipedia_id'] == 43386202]['paragraphs'].iloc[0]

cols = ['kilt_id', 'wikipedia_id', 'wikipedia_title']
idx = pd.MultiIndex.from_tuples(df[cols].values.tolist(), names=cols)


df_p = pd.DataFrame(df['paragraphs'].tolist(), idx).stack().reset_index(cols, name='paragraph')
df_p['paragraph_id'] = df_p.index


df_p

df_p.loc[((df_p['wikipedia_id'] == 339) & (df_p['paragraph_id'] == 100))]

df_p['paragraph'].apply(lambda p: len(p)).describe()

df_p.to_json('../data/kilt_wiki_small_paragraphs.jsonl', orient='records', lines=True)
