==> playground/03_retriev.ipynb <==
from sentence_transformers import SentenceTransformer
import pandas as pd
import numpy as np
import faiss

PARAGRAPHS = '../data/kilt_wiki_small_paragraphs.jsonl'
QUESTIONS = '../data/kilt_fever_train_small.jsonl'
MODEL = 'jinaai/jina-embeddings-v3'
INDEX = '../data/kilt_wiki_small.index'

df_q = pd.read_json(QUESTIONS, lines=True)
df_q

df_p = pd.read_json(PARAGRAPHS, lines=True)
df_p.reset_index()

model = SentenceTransformer(MODEL, trust_remote_code=True)

index = faiss.read_index(INDEX)

def retriev(item, index, k=5):

    task = "retrieval.query"
    query = model.encode(
        item['input'],
        task=task,
        prompt_name=task,
    )

    return index.search(np.array([query]), k)
    
retriev(df_q.iloc[1], index)



df_p.iloc[980].to_dict()
