==> playground/05_create_duckdb.ipynb <==
import duckdb
import pandas as pd

SOURCE = '../data/kilt_wiki_small.duckdb'
WIKI = '../data/kilt_wiki_small.jsonl'
PARAGRAPHS = '../data/kilt_wiki_small_paragraphs.jsonl'

df = pd.read_json(WIKI, orient='records', lines=True)

df_wiki = df[['kilt_id', 'wikipedia_id', 'wikipedia_title', 'text', 'categories']]
df_wiki.head(1)


df_p = pd.read_json(PARAGRAPHS, orient='records', lines=True)

df_p_small = df_p[['kilt_id', 'paragraph_id', 'paragraph']]
df_p_small.head(1)

con = duckdb.connect(SOURCE)

con.execute("SELECT count(*) FROM paragraph;").fetchall()[0][0]

con.sql("DROP TABLE IF EXISTS wiki")
con.sql('CREATE TABLE wiki AS SELECT * FROM df_wiki')
con.sql("INSERT INTO wiki SELECT * FROM df_wiki")

con.sql("DROP TABLE IF EXISTS paragraph")
con.sql("CREATE TABLE paragraph (wikipedia_id VARCHAR, wikipedia_title VARCHAR, id BIGINT, index INTEGER, text VARCHAR);")
con.sql("CREATE SEQUENCE paragraph_id START 1;")

con.execute("""
INSERT INTO paragraph
SELECT
    wikipedia_id,
    wikipedia_title,
    nextval('paragraph_id') as id,
    idx AS index,
    paragraph AS text
FROM wiki, UNNEST(text.paragraph) WITH ORDINALITY AS t(paragraph, idx);
""")

# con.sql('CREATE TABLE paragraph AS SELECT * FROM df_p_small')
# con.sql("INSERT INTO paragraph SELECT * FROM df_p_small")

con.execute("SHOW ALL TABLES").df()

con.execute("SELECT * FROM wiki LIMIT 5").df()

con.execute("SELECT * FROM paragraph LIMIT 5").df()

result = con.execute("SELECT * FROM paragraph LIMIT 5 OFFSET 5").df()
result

BATCH_ROWS = 512
cursor = con.execute("SELECT * FROM paragraph")

rows = cursor.fetchmany(BATCH_ROWS)
rows

def get_paragraph_batch(index, batch_size = 25):
    return con.execute("SELECT * FROM paragraph LIMIT ? OFFSET ?", [batch_size, index*batch_size]).df()


get_paragraph_batch(1)
