==> playground/06_retriev.ipynb <==
from sentence_transformers import SentenceTransformer
import pandas as pd
import numpy as np
import faiss
import duckdb

QUESTIONS = '../data/kilt_fever_train_tiny.jsonl'
MODEL = 'jinaai/jina-embeddings-v3'
INDEX = '../data/kilt_wiki_tiny.index'
SOURCE = '../data/kilt_wiki_tiny.duckdb'

df_q = pd.read_json(QUESTIONS, lines=True)
df_q

model = SentenceTransformer(MODEL, trust_remote_code=True)

index = faiss.read_index(INDEX)

index.is_trained

def retriev(item, index, con, k=5):
    task = "retrieval.query"
    query = model.encode(
        item['input'],
        task=task,
        prompt_name=task,
    )

    distances, indices = index.search(np.array([query]), k)

    # TODO: Why???
    indices = indices+1

    result = con.execute("""
        SELECT *
        FROM paragraph
        WHERE global_id IN ({})
        """.format(",".join(map(str, indices[0])))).df()

    result['d'] = distances[0]

    print(item['input'])
    # print(distances) # TODO: Find out why the distance is always the same?
    return result

con = duckdb.connect(SOURCE)
retriev(df_q.iloc[0], index, con)



con.execute("SELECT * FROM paragraph WHERE wikipedia_id = 43386202").df()

con.execute("""
SELECT * FROM paragraph
WHERE global_id IN ({})
""".format(",".join(map(str, [441])))).df()

res = con.execute("SELECT * FROM paragraph").df()

res['identifier'] = res.apply(lambda row: f"{row['wikipedia_id']}_{row['index']}", axis=1)


ids = res['identifier']
res[ids.isin(ids[ids.duplicated()])].sort_values("identifier")

res


def extract_wikipedia_link(row):

    item = row['output'][0]['provenance'][0]

    return {
        'wikipedia_id': item['wikipedia_id'],
        'section': item['section'],
        'paragraph_id': item['start_paragraph_id'],
        'start_paragraph_id': item['start_paragraph_id'],
        'end_paragraph_id': item['end_paragraph_id']
    }

row = df_q.iloc[0]

ground_truth = extract_wikipedia_link(row)
retrieved_documents = retriev(row, index, con)



retrieved_documents

ground_truth['wikipedia_id'] in retrieved_documents['wikipedia_id'].to_list()

ground_truth['paragraph_id'] in retrieved_documents['index'].to_list()
