==> playground/08_small_fever.ipynb <==
import pandas
import duckdb
from datasets import load_dataset

NAME = 'tiny'
SIZE = 500

WIKI_ALL = f'../data/wikipedia_kilt.duckdb'
OUT_KILT_SUBSET = f'../data/{NAME}_{SIZE}_kilt_fever_train.jsonl'
OUT_WIKI_SUBSET = f'../data/{NAME}_{SIZE}_kilt_wiki.jsonl'

def extract_wikipedia_link(row):

    item = row['output'][0]['provenance'][0]

    return {
        'wikipedia_id': item['wikipedia_id'],
        'section': item['section'],
        'start_paragraph_id': item['start_paragraph_id'],
        'end_paragraph_id': item['end_paragraph_id']
    }

    

kilt_fever = load_dataset("kilt_tasks", name="fever")

kilt_fever = load_dataset("kilt_tasks", name="wiki")

allowed = ['SUPPORTS', 'REFUTES']

train_clean = kilt_fever['train'].filter(lambda row: (row['output'][0]['answer'] in allowed and len(row['output'][0]['provenance']) > 0))
train_clean

subset = train_clean.select(range(SIZE))

relevant = subset.map(extract_wikipedia_link)
relevant_wiki_pages = relevant['wikipedia_id']

relevant.to_json(OUT_KILT_SUBSET)

relevant.to_sql()

relevant_wiki_pages = subset.map(extract_wikipedia_link)['wikipedia_id']

relevant_wiki_pages[0:-1]

con = duckdb.connect(WIKI_ALL)
relevant = con.execute("SELECT * FROM wiki WHERE wikipedia_id in ?", [relevant_wiki_pages[0:-1]]).df()
relevant.to_json(OUT_WIKI_SUBSET, orient='records', lines=True)
