==> playground/09_inspect_db.ipynb <==
import duckdb

DB = '../data/kilt_wiki.duckdb'
WIKI_ALL = '../data/wikipedia_kilt.duckdb'

con = duckdb.connect(DB)

document_id = '23041'
reference_paragraphs = [2, 3, 5]
excluded = [5, 4]

res = con.execute(f"""
    SELECT *
    FROM paragraph
    WHERE wikipedia_id = '{document_id}' AND index IN ({','.join(['?']*len(reference_paragraphs))})
""", reference_paragraphs).df()

res.loc[(~res['index'].isin(excluded) & res['wikipedia_id'] != document_id)].head(5)

con.execute(f"SELECT * FROM paragraph").df()

wiki = duckdb.connect(WIKI_ALL)
wiki.execute(f"""
    SELECT *
    FROM wiki
    WHERE wikipedia_id in ?
""", [['23041', '1']]).df()
