==> playground/11_unanswerable_queries.ipynb <==
import duckdb
from datasets import load_dataset

DB = '../results/2026-01-20_00-46/kilt_wiki.duckdb'
WIKI_ALL = '../data/wikipedia_kilt.duckdb'

# TODO: Collect queries from documents outside of the 10_000 index

con = duckdb.connect(DB)
source = duckdb.connect(WIKI_ALL)

con.execute("SELECT * FROM paragraph").df()

page_ids = con.execute("SELECT wikipedia_id FROM wiki").df()
page_ids = page_ids.values.flatten().tolist()

len(page_ids)

unknown_pages = source.execute(f"SELECT * FROM wiki WHERE wikipedia_id NOT IN ({','.join(['?']*len(page_ids))}) LIMIT 100", page_ids).df()

allowed = ['SUPPORTS', 'REFUTES']

kilt_fever = load_dataset("kilt_tasks", name="fever")
train_clean = kilt_fever['train'].filter(lambda row: (row['output'][0]['answer'] in allowed and len(row['output'][0]['provenance']) > 0))
train_clean

unseen_pages = train_clean.filter(lambda row: row['output'][0]['provenance'][0]['wikipedia_id'] not in page_ids)

unseen_pages[0]

unseen_pages.to_json("../results/validation/unseen.jsonl")
