==> playground/12_results.ipynb <==
import pandas as pd
from omegaconf import OmegaConf
from lib.uncertainty import Simulate
from prettytable import PrettyTable

EXPERIMENTS = [
    '2026-01-26_20-22_fever_dense',
    '2026-01-26_20-23_fever_sparse',
    '2026-01-26_23-23_fever_hybrid',
    '2026-01-26_20-22_fever_similar',
    '2026-01-26_20-25_fever_oracle',
    '2026-01-26_20-55_fever_random',
    '2026-01-26_23-23_fever_probabilistic-random',
    '2026-01-27_17-16_fever_probabilistic-similar',
    '2026-01-26_21-44_compendium_dense',
    '2026-01-26_22-34_compendium_sparse',
    '2026-01-26_23-23_compendium_hybrid',
    '2026-01-26_22-22_compendium_similar',
    '2026-01-26_21-56_compendium_oracle',
    '2026-01-26_22-05_compendium_random',
    '2026-01-26_22-52_compendium_probabilistic-random',
    '2026-01-27_17-17_compendium_probabilistic-similar',
    '2026-01-27_13-07_nq_dense',
    '2026-01-27_17-53_nq_sparse',
    '2026-01-27_18-41_nq_hybrid',
    '2026-01-27_17-51_nq_similar',
    '2026-01-27_17-51_nq_oracle',
    '2026-01-27_17-49_nq_random',
    '2026-01-27_23-07_nq_probabilistic-random',
    '2026-01-27_19-38_nq_probabilistic-similar',
    '2026-01-28_08-57_compendium-de_dense',
    '2026-01-28_09-52_compendium-de_sparse',
    '2026-01-28_10-06_compendium-de_hybrid',
    '2026-01-28_09-35_compendium-de_similar',
    '2026-01-28_09-15_compendium-de_oracle',
    '2026-01-28_09-25_compendium-de_random',
    '2026-01-28_10-29_compendium-de_probabilistic-similar'
]

dataset = {
    'fever': 'Fever',
    'compendium': 'Compendium'
}

strategy = {
    'oracle': 'Oracle',
    'random': 'Random',
    'similar': 'Similar',
    'dense': 'Dense',
    'sparse': 'Sparse'
}

results_all = {}

for experiment in EXPERIMENTS:
    id = experiment.split('_')

    # cfg = OmegaConf.load(f'../results/{experiment}/config.yaml')
    results = pd.read_json(f'../results/{experiment}/results.json')
    results['correct_query'] = True

    results['date'] = id[0]
    results['time'] = id[1]
    results['dataset'] = id[2]

    results['retriever_strategy'] = id[3]

    results_all[experiment] = results

# cfg = OmegaConf.load(f'../results/{experiment}/config.yaml')
simulate = Simulate()

success_rates = {}
for experiment, results in results_all.items():
    success_rates[experiment] = simulate.compute_uncertainty(results)

t = PrettyTable(field_names=['Dataset', 'Retriever Strategy', 'Query', 'Retreiver', 'Generator', '# Tests'])

for experiment, success_rate in success_rates.items():
    id = experiment.split('_')
    t.add_row([
        id[2],
        id[3],
        f"{success_rate['q']['mean']:.2%} ± {success_rate['q']['std']:.4%}",
        f"{success_rate['r']['mean']:.2%} ± {success_rate['r']['std']:.4%}",
        f"{success_rate['g']['mean']:.2%} ± {success_rate['g']['std']:.4%}",
        len(results_all[experiment])
    ]) 

t

df = results_all['2026-01-26_20-22_fever_dense']
df['answer'].value_counts()

df['generated_answer'].value_counts()
