==> playground/12_update_all_db.ipynb <==
import duckdb

TARGET_DB = '../data/all.duckdb'

con = duckdb.connect(TARGET_DB, read_only=True)

# con.sql("ALTER TABLE paragraph ADD COLUMN has_vec BOOLEAN")

# con.sql("UPDATE paragraph SET has_vec = vec IS NOT NULL")

con.execute("SELECT * FROM paragraph WHERE has_vec = TRUE ORDER BY global_id LIMIT 2000").df()

con.execute("SELECT * FROM paragraph WHERE has_vec = TRUE LIMIT 2000").df()

con.execute("SELECT count(*) FROM paragraph WHERE has_vec = FALSE").fetchall()[0][0]

con.execute("SELECT * FROM paragraph WHERE vec IS NOT NULL LIMIT 100").df()
